+++ ExternalConnectors/AWS/EC2FetchDataService.cs
@@ -11,7 +11,7 @@ namespace ExternalConnectors.AWS
         private static List<InstanceInfo>? lastData;
 
         // input must be in format "AWSAPI:instanceid" where instanceid is the ec2 instance id, e.g. i-066f750a76c97583d
-        public static async Task<string> GetEC2InstanceDataAsync(string input)
+        public static async Task<string> GetEC2InstanceDataAsync(string input, string region)
         {
             // get secret id
             if (!input.StartsWith("AWSAPI:"))
@@ -20,19 +20,20 @@ namespace ExternalConnectors.AWS
 
             // init connection credentials, display popup if necessary
             AWSConnectionData.Init();
-            var alldata = await GetEC2IPDataAsync();
+            var alldata = await GetEC2IPDataAsync(region);
             var found = alldata.Where(x => x.InstanceId == InstanceID).SingleOrDefault();
             return (found == null) ? "" : found.PublicIP;
         }
 
-        private static async Task<List<InstanceInfo>> GetEC2IPDataAsync()
+        private static async Task<List<InstanceInfo>> GetEC2IPDataAsync(string region)
         {
             // caching
             TimeSpan timeSpan = DateTime.Now - lastFetch;
             if (timeSpan.TotalMinutes < 1 && lastData != null)
                 return lastData;
 
-            AWSConfigs.AWSRegion = AWSConnectionData.region;
+            //AWSConfigs.AWSRegion = AWSConnectionData.region;
+            AWSConfigs.AWSRegion = region;
             string awsAccessKeyId = AWSConnectionData.awsKeyID;
             string awsSecretAccessKey = AWSConnectionData.awsKey;
 
@@ -82,20 +83,19 @@ namespace ExternalConnectors.AWS
 
             public static string awsKeyID = "";
             public static string awsKey = "";
-            public static string region = "eu-central-1";
+            //public static string _region = "eu-central-1";
 
             public static void Init()
             {
                 if (awsKey != "")
                     return;
-
                 // display gui and ask for data
                 AWSConnectionForm f = new();
                 f.tbAccesKeyID.Text = "" + key.GetValue("KeyID");
                 f.tbAccesKey.Text = "" + key.GetValue("Key");
-                f.tbRegion.Text = "" + key.GetValue("Region");
-                if (f.tbRegion.Text == null || f.tbRegion.Text.Length < 2)
-                    f.tbRegion.Text = region;
+                //f.tbRegion.Text = "" + key.GetValue("Region");
+                //if (f.tbRegion.Text == null || f.tbRegion.Text.Length < 2)
+                //    f.tbRegion.Text = region;
                 _ = f.ShowDialog();
 
                 if (f.DialogResult != DialogResult.OK)
@@ -104,13 +104,13 @@ namespace ExternalConnectors.AWS
                 // store values to memory
                 awsKeyID = f.tbAccesKeyID.Text;
                 awsKey = f.tbAccesKey.Text;
-                region = f.tbRegion.Text;
+                //region = f.tbRegion.Text;
 
 
                 // write values to registry
                 key.SetValue("KeyID", awsKeyID);
                 key.SetValue("Key", awsKey);
-                key.SetValue("Region", region);
+                //key.SetValue("Region", region);
                 key.Close();
             }
         }