+++ mRemoteNG/Connection/ConnectionInitiator.cs
@@ -62,6 +62,18 @@ namespace mRemoteNG.Connection
 
             try
             {
+                if (!string.IsNullOrEmpty(connectionInfo.EC2InstanceId))
+                {
+                    try
+                    {
+                        string host = await ExternalConnectors.AWS.EC2FetchDataService.GetEC2InstanceDataAsync("AWSAPI:" + connectionInfo.EC2InstanceId, connectionInfo.EC2Region);
+                        connectionInfo.Hostname = host;
+                    }
+                    catch
+                    {
+                    }
+                }
+
                 if (connectionInfo.Hostname == "" && connectionInfo.Protocol != ProtocolType.IntApp)
                 {
                     Runtime.MessageCollector.AddMessage(MessageClass.WarningMsg,